# docs/swagger.yaml
basePath: /
definitions:
  github_com_Neimess_shortener_internal_model_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  github_com_Neimess_shortener_internal_model_auth.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_Neimess_shortener_internal_model_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_Neimess_shortener_internal_model_auth.RegisterResponse:
    properties:
      user_id:
        type: integer
    type: object
  github_com_Neimess_shortener_internal_model_url.ShortenRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  github_com_Neimess_shortener_internal_model_url.ShortenResponse:
    properties:
      short:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис для сокращения и редиректа URL.
  title: URL Shortener API
  version: "1.0"
paths:
  /{code}:
    get:
      description: Находит оригинальный URL по короткому коду и делает 302 редирект
      parameters:
      - description: Short URL code
        in: path
        name: code
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "302":
          description: Redirect to original URL
          schema:
            type: string
        "404":
          description: Short code not found
          schema:
            type: string
      summary: Редирект по короткому коду
      tags:
      - urls
  /healthz:
    get:
      description: Возвращает 200 OK, если сервис работает
      produces:
      - text/plain
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Health используется для начального состояния проекта
      tags:
      - health
  /refresh:
    post:
      consumes:
      - application/json
      description: Принимает JSON {"refresh_token":"..."} и возвращает новые access
        и refresh токены
      parameters:
      - description: Refresh Token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_Neimess_shortener_internal_model_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Neimess_shortener_internal_model_auth.RefreshResponse'
        "400":
          description: invalid payload
          schema:
            type: string
        "401":
          description: invalid refresh token
          schema:
            type: string
        "500":
          description: refresh failed
          schema:
            type: string
      summary: Обновление access token
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Принимает JSON {"email":"...","password":"..."} и создаёт нового
        пользователя
      parameters:
      - description: Credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_Neimess_shortener_internal_model_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Neimess_shortener_internal_model_auth.RegisterResponse'
        "400":
          description: invalid payload
          schema:
            type: string
        "500":
          description: failed to shorten url
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
  /shorten:
    post:
      consumes:
      - application/json
      description: 'Принимает JSON {"url": "..."} и возвращает короткий код'
      parameters:
      - description: Original URL
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_Neimess_shortener_internal_model_url.ShortenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Neimess_shortener_internal_model_url.ShortenResponse'
        "400":
          description: invalid payload
          schema:
            type: string
        "500":
          description: failed to shorten url
          schema:
            type: string
      summary: Сокращение URL
      tags:
      - urls
swagger: "2.0"